Refactored common fields and methods to prepare for new schema generation mojo
