+++ metaschema-maven-plugin/src/main/java/gov/nist/secauto/metaschema/maven/plugin/MetaschemaMojo.java
@@ -33,185 +33,38 @@ import gov.nist.secauto.metaschema.model.common.IMetaschema;
 import gov.nist.secauto.metaschema.model.common.MetaschemaException;
 import gov.nist.secauto.metaschema.model.common.util.ObjectUtils;
 
-import org.apache.maven.plugin.AbstractMojo;
-import org.apache.maven.plugin.MojoExecution;
 import org.apache.maven.plugin.MojoExecutionException;
-import org.apache.maven.plugins.annotations.Component;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
-import org.apache.maven.project.MavenProject;
-import org.codehaus.plexus.util.DirectoryScanner;
-import org.sonatype.plexus.build.incremental.BuildContext;
 
 import java.io.File;
 import java.io.IOException;
 import java.io.OutputStream;
-import java.net.URI;
 import java.nio.file.Files;
 import java.nio.file.StandardOpenOption;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import edu.umd.cs.findbugs.annotations.NonNull;
 
 /**
  * Goal which generates Java source files for a given set of Metaschema definitions.
  */
 @Mojo(name = "generate-sources", defaultPhase = LifecyclePhase.PROCESS_SOURCES)
 public class MetaschemaMojo
-    extends AbstractMojo {
-  private static final String SYSTEM_FILE_ENCODING_PROPERTY = "file.encoding";
-  private static final String METASCHEMA_STAE_FILE_NAME = "metaschemaStateFile";
-  private static final String[] DEFAULT_INCLUDES = { "**/*.xml" };
-
-  /**
-   * The Maven project context.
-   * 
-   * @parameter default-value="${project}"
-   * @required
-   * @readonly
-   */
-  @Parameter(defaultValue = "${project}", required = true, readonly = true)
-  MavenProject mavenProject;
-
-  /**
-   * This will be injected if this plugin is executed as part of the standard Maven lifecycle. If the
-   * mojo is directly invoked, this parameter will not be injected.
-   */
-  @Parameter(defaultValue = "${mojoExecution}", readonly = true)
-  private MojoExecution mojoExecution;
-
-  @Component
-  private BuildContext buildContext;
-
-  /**
-   * <p>
-   * The directory where the staleFile is found. The staleFile is used to determine if re-generation
-   * of generated Java classes is needed, by recording when the last build occurred.
-   * </p>
-   * <p>
-   * This directory is expected to be located within the <code>${project.build.directory}</code>, to
-   * ensure that code (re)generation occurs after cleaning the project.
-   * </p>
-   */
-  @Parameter(defaultValue = "${project.build.directory}/metaschema", readonly = true, required = true)
-  protected File staleFileDirectory;
-
-  /**
-   * <p>
-   * Defines the encoding used for generating Java Source files.
-   * </p>
-   * <p>
-   * The algorithm for finding the encoding to use is as follows (where the first non-null value found
-   * is used for encoding):
-   * <ol>
-   * <li>If the configuration property is explicitly given within the plugin's configuration, use that
-   * value.</li>
-   * <li>If the Maven property <code>project.build.sourceEncoding</code> is defined, use its
-   * value.</li>
-   * <li>Otherwise use the value from the system property <code>file.encoding</code>.</li>
-   * </ol>
-   * </p>
-   * 
-   * @see #getEncoding(boolean)
-   * @since 2.0
-   */
-  @Parameter(defaultValue = "${project.build.sourceEncoding}")
-  private String encoding;
-
-  /**
-   * Location to generate Java source files in.
-   */
-  @Parameter(defaultValue = "${project.build.directory}/generated-sources/metaschema", required = true)
-  private File outputDirectory;
-
-  /**
-   * The directory to read source metaschema from.
-   */
-  @Parameter(defaultValue = "${basedir}/src/main/metaschema")
-  private File metaschemaDir;
-
-  /**
-   * A set of inclusion patterns used to select which metaschema are to be processed. By default, all
-   * files are processed.
-   */
-  @Parameter
-  protected String[] includes;
-
-  /**
-   * A set of exclusion patterns used to prevent certain files from being processed. By default, this
-   * set is empty such that no files are excluded.
-   */
-  @Parameter
-  protected String[] excludes;
-
-  /**
-   * Indicate if the execution should be skipped.
-   */
-  @Parameter(property = "metaschema.skip", defaultValue = "false")
-  private boolean skip;
+    extends AbstractMetaschemaMojo {
+  private static final String METASCHEMA_STALE_FILE_NAME = "metaschemaStaleFile";
 
   /**
    * A set of binding configurations.
    */
   @Parameter
   protected File[] configs;
-
-  /**
-   * The BuildContext is used to identify which files or directories were modified since last build.
-   * This is used to determine if java code generation must be performed again.
-   *
-   * @return the active Plexus BuildContext.
-   */
-  protected final BuildContext getBuildContext() {
-    return buildContext;
-  }
-
-  /**
-   * Retrieve the Maven project context.
-   * 
-   * @return The active MavenProject.
-   */
-  protected final MavenProject getMavenProject() {
-    return mavenProject;
-  }
-
-  /**
-   * Retrieve the mojo execution context.
-   * 
-   * @return The active MojoExecution.
-   */
-  @SuppressFBWarnings(value = "EI_EXPOSE_REP", justification = "this is a data holder")
-  public MojoExecution getMojoExecution() {
-    return mojoExecution;
-  }
-
-  /**
-   * Retrieve the directory where generated classes will be stored.
-   * 
-   * @return the directory
-   */
-  protected File getOutputDirectory() {
-    return outputDirectory;
-  }
-
-  /**
-   * Set the directory where generated classes will be stored.
-   * 
-   * @param outputDirectory
-   *          the directory to use
-   */
-  protected void setOutputDirectory(File outputDirectory) {
-    Objects.requireNonNull(outputDirectory, "outputDirectory");
-    this.outputDirectory = outputDirectory;
-  }
 
   /**
    * <p>
@@ -223,68 +76,9 @@ public class MetaschemaMojo
    *
    * @return the stale filename postfix
    */
+  @Override
   protected String getStaleFileName() {
-    return METASCHEMA_STAE_FILE_NAME;
-  }
-
-  /**
-   * Gets the staleFile for this execution.
-   *
-   * @return the staleFile
-   */
-  protected final File getStaleFile() {
-    StringBuilder builder = new StringBuilder();
-    if (getMojoExecution() != null) {
-      builder.append(getMojoExecution().getExecutionId()).append('-');
-    }
-    builder.append(getStaleFileName());
-    return new File(staleFileDirectory, builder.toString());
-  }
-
-  /**
-   * Gets the file encoding to use for generated classes.
-   * <p>
-   * The algorithm for finding the encoding to use is as follows (where the first non-null value found
-   * is used for encoding):
-   * </p>
-   * <ol>
-   * <li>If the configuration property is explicitly given within the plugin's configuration, use that
-   * value.</li>
-   * <li>If the Maven property <code>project.build.sourceEncoding</code> is defined, use its
-   * value.</li>
-   * <li>Otherwise use the value from the system property <code>file.encoding</code>.</li>
-   * </ol>
-   *
-   * @return The encoding to be used by this AbstractJaxbMojo and its tools.
-   */
-  protected final String getEncoding() {
-    String encoding;
-    if (this.encoding != null) {
-      // first try to use the provided encoding
-      encoding = this.encoding;
-      getLog().debug(String.format("Using configured encoding [%s].", encoding));
-    } else {
-      encoding = System.getProperty(SYSTEM_FILE_ENCODING_PROPERTY);
-      getLog().warn(String.format("Using system encoding [%s]. This build is platform dependent!", encoding));
-    }
-    return encoding;
-  }
-
-  /**
-   * Retrieve a stream of Metaschema file sources.
-   * 
-   * @return the stream
-   */
-  protected Stream<File> getSources() {
-    DirectoryScanner ds = new DirectoryScanner();
-    ds.setBasedir(metaschemaDir);
-    ds.setIncludes(includes != null && includes.length > 0 ? includes : DEFAULT_INCLUDES);
-    ds.setExcludes(excludes != null && excludes.length > 0 ? excludes : null);
-    ds.addDefaultExcludes();
-    ds.setCaseSensitive(true);
-    ds.setFollowSymlinks(false);
-    ds.scan();
-    return Stream.of(ds.getIncludedFiles()).map(filename -> new File(metaschemaDir, filename)).distinct();
+    return METASCHEMA_STALE_FILE_NAME;
   }
 
   /**
@@ -302,52 +96,27 @@ public class MetaschemaMojo
     return retval;
   }
 
-  /**
-   * Determine if the execution of this mojo should be skipped.
-   * 
-   * @return {@code true} if the mojo execution should be skipped, or {@code false} otherwise
-   */
-  protected boolean shouldExecutionBeSkipped() {
-    return skip;
-  }
-
-  /**
-   * Determine if code generation is required. This is done by comparing the last modified time of
-   * each Metaschema source file against the stale file managed by this plugin.
-   * 
-   * @return {@code true} if the code generation is needed, or {@code false} otherwise
-   */
-  protected boolean isGenerationRequired() {
-    final File staleFile = getStaleFile();
-    boolean generate = !staleFile.exists();
-    if (generate) {
-      getLog().info(String.format("Stale file '%s' doesn't exist! Generating source files.", staleFile.getPath()));
-      generate = true;
-    } else {
-      generate = false;
-      // check for staleness
-      long staleLastModified = staleFile.lastModified();
-
-      BuildContext buildContext = getBuildContext();
-      URI metaschemaDirRelative = getMavenProject().getBasedir().toURI().relativize(metaschemaDir.toURI());
-
-      if (buildContext.isIncremental() && buildContext.hasDelta(metaschemaDirRelative.toString())) {
-        getLog().info("metaschemaDirRelative: " + metaschemaDirRelative.toString());
-        generate = true;
-      }
-
-      if (!generate) {
-        for (File sourceFile : getSources().collect(Collectors.toList())) {
-          getLog().info("Source file: " + sourceFile.getPath());
-          if (sourceFile.lastModified() > staleLastModified) {
-            generate = true;
-          }
-        }
+  protected void generate(@NonNull Set<IMetaschema> metaschemaCollection) throws MojoExecutionException {
+    DefaultBindingConfiguration bindingConfiguration = new DefaultBindingConfiguration();
+    for (File config : getConfigs()) {
+      try {
+        getLog().info("Loading binding configuration: " + config.getPath());
+        bindingConfiguration.load(config);
+      } catch (IOException ex) {
+        throw new MojoExecutionException(
+            String.format("Unable to load binding configuration from '%s'.", config.getPath()), ex);
       }
     }
-    return generate;
-  }
 
+    try {
+      getLog().info("Generating Java classes in: " + getOutputDirectory().getPath());
+      JavaGenerator.generate(metaschemaCollection, ObjectUtils.notNull(getOutputDirectory().toPath()),
+          bindingConfiguration);
+    } catch (IOException ex) {
+      throw new MojoExecutionException("Creation of Java classes failed.", ex);
+    }
+  }
+  
   @Override
   public void execute() throws MojoExecutionException {
     File staleFile = getStaleFile();
@@ -379,6 +148,7 @@ public class MetaschemaMojo
         }
       }
 
+      
       // generate Java sources based on provided metaschema sources
       final MetaschemaLoader loader = new MetaschemaLoader();
       loader.allowEntityResolution();
@@ -394,25 +164,7 @@ public class MetaschemaMojo
         metaschemaCollection.add(metaschema);
       }
 
-      // TODO: load this from the requested file
-      DefaultBindingConfiguration bindingConfiguration = new DefaultBindingConfiguration();
-      for (File config : getConfigs()) {
-        try {
-          getLog().info("Loading binding configuration: " + config.getPath());
-          bindingConfiguration.load(config);
-        } catch (IOException ex) {
-          throw new MojoExecutionException(
-              String.format("Unable to load binding configuration from '%s'.", config.getPath()), ex);
-        }
-      }
-
-      try {
-        getLog().info("Generating Java classes in: " + getOutputDirectory().getPath());
-        JavaGenerator.generate(metaschemaCollection, ObjectUtils.notNull(getOutputDirectory().toPath()),
-            bindingConfiguration);
-      } catch (IOException ex) {
-        throw new MojoExecutionException("Creation of Java classes failed.", ex);
-      }
+      generate(metaschemaCollection);
 
       // create the stale file
       if (!staleFileDirectory.exists()) {
@@ -430,7 +182,7 @@ public class MetaschemaMojo
       }
 
       // for m2e
-      buildContext.refresh(getOutputDirectory());
+      getBuildContext().refresh(getOutputDirectory());
     }
 
     // add generated sources to Maven
